feat: change clickArea all tag from button to accessible div with role button
We want to be able to nest a button in the header of the accordian item, and we cant nest two buttons.
This commit accessibly changes the role of the header div to a button when the clickarea is "all".
This includes adding a tab index, keyboard listeners, role, and click handler, all conditionally spread onto the div.