feat: hide a tooltip by ID via a new method wrapping a custom document event
There's really no way to trigger a tooltip being shown or hidden instead of hovering or clicking
a trigger element. We can use forceShow and HideTooltip, but we lose out on the trigger element
and animations. This commit adds a method for hiding a shown tooltip via a custom JS event.

We will use this to close the contextMenu onClick of any of the items.

This required dispatching state and removing click listeners in the case of useClickInsteadOfHover.