feat: add showTooltip helper function, add new props, and move Portal inside target div
Adds separate showTooltip helper similar to hideTooltip. Works for hover and click tooltips, works best for click ones.
Adds a story as well. Removes some of the code outside of the detect hook, since it wasn't necessary.

Moves the Portal inside of the target container, so that we can capture popper events via the target container. This is
important in ContextMenu keyboard nav. It won't break any existing listeners in Core - none of the tooltips really have any
listeners on them.

Adds noAnimations prop to Tooltip, which changes the transition duration to 1ms. We can't do zero because we really on transitionEnd events.

Also adds a stopPopperClickPropagation, to stop clicks on the popper from propagating out. To do this without accessing
the NativeEvent, the document event listeners had to be changed to window listeners, as React bubbles everything up to
document before doing events.

Also removes the useForwardedRef call from Tooltip, since it was unecessary and didn't work to pass into combinedRef.