fix: remove remove default prevention from buttonBase and style based on focus-visible
We were preventing default because I didn't want focus styles in storybook, but really we could just
style based off of focus-visible. This changes that.

Removes useUsingMouse helper too since we don't need it after these changes.

Also focuses the threedotbutton when closing the contextmenu to restore the tab order.