fix: focus trigger in the onClick of menuItems and when hitting tab or focus
This fixes a race condition that was happening when focusing the threedot button onHide of the tooltip.
It was happening too late and grabbing focus, which was an issue for buttons in Core that grab focus, such as
popping modals, creating a new group, etc. Moving up the focus fixes this.